fix(ui): wait out the summarizer before a static autorun advance
