+++ apps/desktop/src/store/slices/workflows/orchestrateNextStep.ts
@@ -54,6 +54,7 @@ import { preSpawnWorkflowAgents } from './preSpawnWorkflowAgents';
 import { patchWorkflowRun, withoutKeys } from './patchWorkflowRun';
 import { recordOrchestratorUsage } from './recordOrchestratorUsage';
 import { findWorkflowActivationBlock } from './workflowActivationGate';
+import { waitForSessionSummarizer } from './summarizerGate';
 import { WORKFLOW_BLOCK_COPY } from '../../../features/workflows/blockCopy';
 import type { GetFn, SetFn } from './types';
 
@@ -64,29 +65,6 @@ export type OrchestrateOptions = {
 };
 
 const orchestrationInFlight = new Set<WorkflowRunId>();
-
-const SUMMARIZER_GATE_POLL_MS = 100;
-const SUMMARIZER_GATE_TIMEOUT_MS = 60_000;
-
-type SummarizerGateParams = {
-  readonly get: GetFn;
-  readonly sessionId: SessionId;
-};
-
-const waitForSessionSummarizer = async ({
-  get,
-  sessionId,
-}: SummarizerGateParams): Promise<void> => {
-  const deadline = Date.now() + SUMMARIZER_GATE_TIMEOUT_MS;
-  while (get().summarizerStatus[sessionId]?.status === 'running') {
-    if (Date.now() >= deadline) {
-      return;
-    }
-    await new Promise<void>((resolve) => {
-      setTimeout(resolve, SUMMARIZER_GATE_POLL_MS);
-    });
-  }
-};
 
 type DecidingParams = {
   readonly set: SetFn;
